+++ Servo/Servo_SPT5535LV360.cpp
@@ -1,6 +1,6 @@
 #include "Servo_SPT5535LV360.h"
 
-Servo_SPT5535LV360::Servo_SPT5535LV360(int pin) : m_pin(pin) {
+Servo_SPT5535LV360::Servo_SPT5535LV360(uint8_t pin) : m_pin(pin) {
     softPwmCreate(m_pin, 0, 100); // Создание PWM на пине для управления сервоприводом
 }
